Fix for platfrom detection
