+++ lib/train-k8s-container/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module TrainPlugins
+  module K8sContainer
+    VERSION = "1.3.0"
+  end
+end
