--- lib/train/k8s/container/version.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Train
-  module K8s
-    module Container
-      VERSION = "1.3.0"
-    end
-  end
-end
